+++ benchmarks/bm2d/bm2d_cpu_omp_kernel.c
@@ -1,7 +1,8 @@
 #include <homp.h>
 #include "bm2d.h"
 
-void bm2d_cpu_omp_wrapper(omp_offloading_t *off, int start, int len, long n, long m, int u_dimX, int u_dimY, REAL *u, REAL *uold, int maxwin, int coeff_dimX, REAL *coeff)
+void bm2d_cpu_omp_wrapper(omp_offloading_t *off, long start, long len, long n, long m, long u_dimX, long u_dimY,
+                          REAL *u, REAL *uold, int maxwin, int coeff_dimX, REAL *coeff)
 {
 #if CORRECTNESS_CHECK
 	    	BEGIN_SERIALIZED_PRINTF(off->devseqid);
