+++ benchmarks/bm2d/bm2d_nvgpu_cuda_kernel.cu
@@ -76,7 +76,7 @@ __global__ void bm2d_nvgpu_kernel(int start_n, int len_n, long n, long m, int u_
     }
 }
 
-void bm2d_nvgpu_cuda_wrapper(omp_offloading_t *off, int start, int len, long n, long m, int u_dimX, int u_dimY, REAL *u, REAL *uold, int maxwin, int coeff_dimX, REAL *coeff) {
+void bm2d_nvgpu_cuda_wrapper(omp_offloading_t *off, long start, long len, long n, long m, long u_dimX, long u_dimY, REAL *u, REAL *uold, int maxwin, int coeff_dimX, REAL *coeff) {
     dim3 threads_per_team(16, 16);
     dim3 teams_per_league((len + threads_per_team.x - 1) / threads_per_team.x,
                           (m + threads_per_team.y - 1) / threads_per_team.y); /* we assume dividable */
